Move DataBuf to own file and enable lints
